python/Backend/pruebas_fastapi/routers/jwt_auth_users.py
# Users API con autorización OAuth2 JWT

from fastapi import APIRouter, Depends, HTTPException, status
#from fastapi import FastAPI, Depends, HTTPException, status
from pydantic import BaseModel
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from jose import jwt, JWTError
from passlib.context import CryptContext
from datetime import datetime, timedelta

ALGORITHM = "HS256" # algoritmo de incriptacion
ACCESS_TOKEN_DURATION = 1 # para definir el tiempo de la autenticacion
SECRET = "201d573bd7d1344d3a3bfce1550b69102fd11be3db6d379508b6cccc58ea230b" # nos ayuda a darle una capa extra de seguridad a la encriptacion de nuestro access_token

router = APIRouter(prefix="/jwtauth",
                   tags=["jwtauth"],
                   responses={status.HTTP_404_NOT_FOUND: {"message": "No encontrado"}})

#app=FastAPI()
oauth2 = OAuth2PasswordBearer(tokenUrl="login")

crypt = CryptContext(schemes=["bcrypt"])


class User(BaseModel):
    username: str
    full_name: str
    email: str
    disabled: bool


class UserDB(User):
    password: str


users_db = {
    "joseisaac": {
        "username": "joseisaac",
        "full_name": "jose isaac",
        "email": "joseisaac@example.com",
        "disabled": False,
        "password": "$2a$12$u9We2RbG6cv2JeOwahBOoOhbwS3OGc9TA7z.VMRuFpvDzUk8SHAE2"
    },
    "joseisaac2": {
        "username": "joseisaac2",
        "full_name": "jose isaac 2",
        "email": "joseisaac2@example.com",
        "disabled": True,
        "password": "$2a$12$x/D4CkEC5AHR.dblhgMBueh7Lg4JW7nTflpIJtQ9soc1YpIpJykuG"
    }
}


def search_user_db(username: str):
    if username in users_db:
        return UserDB(**users_db[username])


def search_user(username: str):
    if username in users_db:
        return User(**users_db[username])


async def auth_user(token: str = Depends(oauth2)):

    exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Credenciales de autenticación inválidas",
        headers={"WWW-Authenticate": "Bearer"})

    try:
        username = jwt.decode(token, SECRET, algorithms=[ALGORITHM]).get("sub")
        if username is None:
            raise exception

    except JWTError:
        raise exception

    return search_user(username)


async def current_user(user: User = Depends(auth_user)):
    if user.disabled:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Usuario inactivo")

    return user


@router.post("/login")
async def login(form: OAuth2PasswordRequestForm = Depends()):

    user_db = users_db.get(form.username)
    if not user_db:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST, detail="El usuario no es correcto")

    user = search_user_db(form.username)

    if not crypt.verify(form.password, user.password):
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST, detail="La contraseña no es correcta")
    # El access_token lo definimos como un json para luego encriptarlo(jsonwebtoken=jwt)
    access_token = {"sub": user.username,
                    "exp": datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_DURATION)}

    return {"access_token": jwt.encode(access_token, SECRET, algorithm=ALGORITHM), "token_type": "bearer"}


@router.get("/users/me")
async def me(user: User = Depends(current_user)):
    return user
