python/Backend/pruebas_fastapi/requirements.txt
# lista de todas dependecias que se instalaron en el proyecto para garantizar de que se instalen en el servidor remoto
fastapi
python-jose
passlib
bcrypt
pymongo
python-multipart